WIP: refactor(table): refactor the test cases (#2232)
* fix(table): move common test util out of table

* refactor(table): move column test out of table

* refactor(table): remove unnecessary toArray

* refactor(table): use triggerEvent from test-util

* fix(utils): add types to triggerEvent

* fix(table): add commented test back

* refactor(table): remove done from async test

* refactor(table): remove unused expectArray

* fix(table): fix type error on triggerEvent

* refactor(table): use conciser array methods
